+++ src/binders.js
@@ -1,8 +1,46 @@
-export const binders ={
+import { entryObject } from './utils/utils.js';
+
+export const binders = {
   productContainerBinder: null,
   productInventoryContainerBinder: null,
 };
 
 export function render(binderName) {
   binders[binderName]?.();
+}
+
+export function createBinder(view, viewModelCreator) {
+  const render = () => {
+    const vm = viewModelCreator();
+      const el = bindViewModelWithView(vm, view);
+      view.render(el);
+  }
+  render();
+
+  return render;
+}
+
+function bindViewModelWithView(vm, view) {
+  entryObject(vm).forEach(([elementName, elementViewModel]) => {
+    const pairElement = view[elementName];
+    entryObject(elementViewModel).forEach(([key, val]) => {
+      if (key === 'events') {
+        entryObject(val).forEach(([eventType, callback]) => {
+          pairElement.addEventListener(eventType, callback);
+        });
+        return;
+      }
+
+      if (key === 'ref') {
+        val.element = pairElement;
+        return;
+      }
+
+      if (key === 'children') {
+        pairElement.innerHTML = val;
+      }
+    });
+  });
+
+  return view.rootElement;
 }
